chore: rename modules package
